% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseRefSet.R
\name{chooseRefSet}
\alias{chooseRefSet}
\title{A function to choose a data-driven reference set.}
\usage{
chooseRefSet(fitted_model, reference_set_size = 50)
}
\arguments{
\item{fitted_model}{The output from \code{radEmu::emuFit} when applied to data with a
constraint over all taxa.}

\item{reference_set_size}{The size of the reference set if it is data-driven, default
is set to \code{50}. We recommend a reference set of size 30-100 for the best balance
of computational efficiency and estimation precision.}
}
\value{
The set of taxa of size \code{reference_set_size} with the smallest absolute
estimated log fold-differences for the covariate, relative to the smoothed median
log fold-difference over all taxa. Also a vector of length p (the number of columns
in the design matrix) of differences between the smoothed median log fold-difference over
all taxa and the smoothed median log fold-difference over the reference set.
}
\description{
A function to choose a data-driven reference set.
}
